snappy,systemd: move GenServiceFile, GetSocketFile from systemd/
This patch moves functions that take ServiceDescription to services.go.
The rationale is that ServiceDescription will decay to little (if
anything) more than just snap.AppInfo and to avoid circular dependencies
we don't want to import snap.Info from systemd/. In addition this has a
nice feature of being symmetric with binaries.go

The code can be refactored further, there's some duplication in
services_test.go and ServiceDescription is now mostly a useless
pass-throuh. Those will be addressed by subsequent branches.
