+++ systemd/systemd.go
@@ -29,15 +29,10 @@ import (
 	"path/filepath"
 	"regexp"
 	"strconv"
-	"strings"
-	"text/template"
 	"time"
 
-	"github.com/ubuntu-core/snappy/arch"
 	"github.com/ubuntu-core/snappy/dirs"
-	"github.com/ubuntu-core/snappy/logger"
 	"github.com/ubuntu-core/snappy/osutil"
-	"github.com/ubuntu-core/snappy/snap/snapenv"
 )
 
 var (
@@ -93,8 +88,6 @@ type Systemd interface {
 	Stop(service string, timeout time.Duration) error
 	Kill(service, signal string) error
 	Restart(service string, timeout time.Duration) error
-	GenServiceFile(desc *ServiceDescription) string
-	GenSocketFile(desc *ServiceDescription) string
 	Status(service string) (string, error)
 	ServiceStatus(service string) (*ServiceStatus, error)
 	Logs(services []string) ([]Log, error)
@@ -117,7 +110,7 @@ const (
 	RestartAlways     RestartCondition = "always"
 )
 
-var restartMap = map[string]RestartCondition{
+var RestartMap = map[string]RestartCondition{
 	"never":       RestartNever,
 	"on-success":  RestartOnSuccess,
 	"on-failure":  RestartOnFailure,
@@ -141,7 +134,7 @@ func (rc *RestartCondition) UnmarshalYAML(unmarshal func(interface{}) error) err
 		return err
 	}
 
-	nrc, ok := restartMap[v]
+	nrc, ok := RestartMap[v]
 	if !ok {
 		return ErrUnknownRestartCondition
 	}
@@ -151,37 +144,7 @@ func (rc *RestartCondition) UnmarshalYAML(unmarshal func(interface{}) error) err
 	return nil
 }
 
-// ServiceDescription describes a snappy systemd service
-type ServiceDescription struct {
-	SnapName        string
-	AppName         string
-	Version         string
-	Revision        int
-	Description     string
-	SnapPath        string
-	Start           string
-	Stop            string
-	PostStop        string
-	StopTimeout     time.Duration
-	Restart         RestartCondition
-	Type            string
-	AaProfile       string
-	BusName         string
-	UdevAppName     string
-	Socket          bool
-	SocketFileName  string
-	ListenStream    string
-	SocketMode      string
-	ServiceFileName string
-}
-
 const (
-	// the default target for systemd units that we generate
-	servicesSystemdTarget = "multi-user.target"
-
-	// the default target for systemd units that we generate
-	socketsSystemdTarget = "sockets.target"
-
 	// the location to put system services
 	snapServicesDir = "/etc/systemd/system"
 )
@@ -332,115 +295,6 @@ func (s *systemd) Stop(serviceName string, timeout time.Duration) error {
 	}
 
 	return &Timeout{action: "stop", service: serviceName}
-}
-
-func (s *systemd) GenServiceFile(desc *ServiceDescription) string {
-	serviceTemplate := `[Unit]
-Description={{.Description}}
-After=snapd.frameworks.target{{ if .Socket }} {{.SocketFileName}}{{end}}
-Requires=snapd.frameworks.target{{ if .Socket }} {{.SocketFileName}}{{end}}
-X-Snappy=yes
-
-[Service]
-ExecStart=/usr/bin/ubuntu-core-launcher {{.UdevAppName}} {{.AaProfile}} {{.FullPathStart}}
-Restart={{.Restart}}
-WorkingDirectory=/var{{.SnapPath}}
-Environment={{.EnvVars}}
-{{if .Stop}}ExecStop=/usr/bin/ubuntu-core-launcher {{.UdevAppName}} {{.AaProfile}} {{.FullPathStop}}{{end}}
-{{if .PostStop}}ExecStopPost=/usr/bin/ubuntu-core-launcher {{.UdevAppName}} {{.AaProfile}} {{.FullPathPostStop}}{{end}}
-{{if .StopTimeout}}TimeoutStopSec={{.StopTimeout.Seconds}}{{end}}
-Type={{.Type}}
-{{if .BusName}}BusName={{.BusName}}{{end}}
-
-[Install]
-WantedBy={{.ServiceSystemdTarget}}
-`
-	var templateOut bytes.Buffer
-	t := template.Must(template.New("wrapper").Parse(serviceTemplate))
-
-	restartCond := desc.Restart.String()
-	if restartCond == "" {
-		restartCond = RestartOnFailure.String()
-	}
-
-	wrapperData := struct {
-		// the service description
-		ServiceDescription
-		// and some composed values
-		FullPathStart        string
-		FullPathStop         string
-		FullPathPostStop     string
-		ServiceSystemdTarget string
-		SnapArch             string
-		Home                 string
-		EnvVars              string
-		SocketFileName       string
-		Restart              string
-		Type                 string
-	}{
-		*desc,
-		filepath.Join(desc.SnapPath, desc.Start),
-		filepath.Join(desc.SnapPath, desc.Stop),
-		filepath.Join(desc.SnapPath, desc.PostStop),
-		servicesSystemdTarget,
-		arch.UbuntuArchitecture(),
-		// systemd runs as PID 1 so %h will not work.
-		"/root",
-		"",
-		desc.SocketFileName,
-		restartCond,
-		desc.Type,
-	}
-	allVars := snapenv.GetBasicSnapEnvVars(wrapperData)
-	allVars = append(allVars, snapenv.GetUserSnapEnvVars(wrapperData)...)
-	wrapperData.EnvVars = "\"" + strings.Join(allVars, "\" \"") + "\"" // allVars won't be empty
-
-	if err := t.Execute(&templateOut, wrapperData); err != nil {
-		// this can never happen, except we forget a variable
-		logger.Panicf("Unable to execute template: %v", err)
-	}
-
-	return templateOut.String()
-}
-
-func (s *systemd) GenSocketFile(desc *ServiceDescription) string {
-	serviceTemplate := `[Unit]
-Description={{.Description}} Socket Unit File
-PartOf={{.ServiceFileName}}
-X-Snappy=yes
-
-[Socket]
-ListenStream={{.ListenStream}}
-{{if .SocketMode}}SocketMode={{.SocketMode}}{{end}}
-
-[Install]
-WantedBy={{.SocketSystemdTarget}}
-`
-	var templateOut bytes.Buffer
-	t := template.Must(template.New("wrapper").Parse(serviceTemplate))
-
-	wrapperData := struct {
-		// the service description
-		ServiceDescription
-		// and some composed values
-		ServiceFileName,
-		ListenStream string
-		SocketMode          string
-		SocketSystemdTarget string
-	}{
-		*desc,
-		desc.ServiceFileName,
-		desc.ListenStream,
-		desc.SocketMode,
-		socketsSystemdTarget,
-	}
-
-	if err := t.Execute(&templateOut, wrapperData); err != nil {
-		// this can never happen, except we forget a variable
-		logger.Panicf("Unable to execute template: %v", err)
-	}
-
-	return templateOut.String()
 }
 
 // Kill all processes of the unit with the given signal
